move NotImportedModule to utils/utils.py
