+++ darts/tests/models/forecasting/test_probabilistic_models.py
@@ -16,11 +16,11 @@ from darts.models import (
     ExponentialSmoothing,
     LightGBMModel,
     LinearRegressionModel,
-    NotImportedModule,
     XGBModel,
 )
 from darts.tests.conftest import TORCH_AVAILABLE, tfm_kwargs
 from darts.utils import timeseries_generation as tg
+from darts.utils.utils import NotImportedModule
 
 logger = get_logger(__name__)
